Move the directory of a renderer for Spine 2.1
